CLN refactor the win32 spawning main
