+++ loky/backend/popen_loky_win32.py
@@ -1,10 +1,11 @@
 import os
 import sys
+from pickle import load
+from multiprocessing import process, util
 
-from .context import get_spawning_popen, set_spawning_popen
 from . import spawn
 from . import reduction
-from multiprocessing import util
+from .context import get_spawning_popen, set_spawning_popen
 
 if sys.platform == "win32":
     # Avoid import error by code introspection tools such as test runners
@@ -50,8 +51,7 @@ class Popen(_Popen):
         if sys.version_info[:2] > (3, 3):
             wfd = msvcrt.open_osfhandle(wfd, 0)
 
-        cmd = spawn.get_command_line(parent_pid=os.getpid(),
-                                     pipe_handle=rhandle)
+        cmd = get_command_line(parent_pid=os.getpid(), pipe_handle=rhandle)
         cmd = ' '.join('"%s"' % x for x in cmd)
 
         try:
@@ -64,9 +64,8 @@ class Popen(_Popen):
                         None, None, inherit, 0,
                         None, None, None)
                     _winapi.CloseHandle(ht)
-                except:
+                finally:
                     _winapi.CloseHandle(rhandle)
-                    raise
 
                 # set attributes of self
                 self.pid = pid
@@ -97,3 +96,54 @@ class Popen(_Popen):
     def duplicate_for_child(self, handle):
         assert self is get_spawning_popen()
         return reduction.duplicate(handle, self.sentinel)
+
+
+if sys.version_info[:2] >= (3, 4):
+    from multiprocessing.spawn import get_command_line
+else:
+    # compatibility for python2.7. Duplicate here the code from
+    # multiprocessing.forking.main to call our prepare function and correctly
+    # set the default start_methods in loky.
+
+    def get_command_line(pipe_handle, **kwds):
+        '''
+        Returns prefix of command line used for spawning a child process
+        '''
+        if getattr(sys, 'frozen', False):
+            return ([sys.executable, '--multiprocessing-fork', pipe_handle])
+        else:
+            prog = 'from loky.backend.popen_loky_win32 import main; main()'
+            opts = util._args_from_interpreter_flags()
+            return [spawn.get_executable()] + opts + [
+                '-c', prog, '--multiprocessing-fork', pipe_handle]
+
+    def is_forking(argv):
+        '''
+        Return whether commandline indicates we are forking
+        '''
+        if len(argv) >= 2 and argv[1] == '--multiprocessing-fork':
+            assert len(argv) == 3
+            return True
+        else:
+            return False
+
+    def main():
+        '''
+        Run code specified by data received over pipe
+        '''
+        assert is_forking(sys.argv)
+
+        handle = int(sys.argv[-1])
+        fd = msvcrt.open_osfhandle(handle, os.O_RDONLY)
+        from_parent = os.fdopen(fd, 'rb')
+
+        process.current_process()._inheriting = True
+        preparation_data = load(from_parent)
+        spawn.prepare(preparation_data)
+        self = load(from_parent)
+        process.current_process()._inheriting = False
+
+        from_parent.close()
+
+        exitcode = self._bootstrap()
+        exit(exitcode)
